+++ projects/cloud/app/models/organization.rb
@@ -5,6 +5,7 @@ class Organization < ApplicationRecord
 
   # Associations
   has_one :account, as: :owner, class_name: "Account", dependent: :destroy
+  has_many :invitations, as: :organization, dependent: :destroy
   # Inspired from: https://github.com/RolifyCommunity/rolify/wiki/Usage#finding-roles-through-associations
   has_many :users, -> {
  where(roles: { name: :user }).where.not(roles: { name: :admin }) }, through: :roles, class_name: "User", source: :users